[CF-30] Refactor and remove confluent specific configs
